agent/requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn>=0.20.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=0.19.0
httpx>=0.23.0
python-multipart>=0.0.5